# swagger.yaml
consumes:
- application/json
...
host: http://api.recipes.io:44004
info:
  contact:
    email: jersonsatoru@yahoo.com.br
  description: |-
    Recipes API
    This is a simple recipes API. You can find out more about the API at  google, thanks
  version: 1.0.0
paths:
...
  /refresh:
    post:
      description: Refresh JWT Token
      operationId: refreshToken
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
        "401":
          description: Invalid credentials
      tags:
      - token
...
produces:
- application/json
schemes:
- https
securityDefinitions:
  api_key:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"